requirements.txt
numpy>=1.24
Pillow>=10.0
scipy>=1.10
matplotlib>=3.7